chore: move fs functions to own module
